Renovation: jest tests refactoring (#13811)
* Renovation: jest tests refactoring

* fix tests
