+++ js/renovation/__tests__/jquery/test_components/templated_test_widget.tsx
