+++ js/renovation/button.tsx
@@ -16,10 +16,10 @@ import noop from './utils/noop';
 import * as themes from '../ui/themes';
 import { click } from '../events/short';
 import { getImageSourceType } from '../core/utils/icon';
-import { Icon } from './icon';
-import { InkRipple } from './ink-ripple';
-import { Widget } from './widget';
-import BaseWidgetProps from './utils/base-props';
+import { Icon } from './common/icon';
+import { InkRipple } from './common/ink_ripple';
+import { Widget } from './common/widget';
+import BaseWidgetProps from './utils/base_props';
 import BaseComponent from './preact_wrapper/button';
 
 const stylingModes = ['outlined', 'text', 'contained'];
