+++ js/renovation/common/__tests__/icon.test.tsx
@@ -1,6 +1,6 @@
 import { h } from 'preact';
 import { shallow } from 'enzyme';
-import { Icon } from '../../js/renovation/icon';
+import { Icon } from '../icon';
 
 describe('Icon', () => {
   describe('Props', () => {
