+++ js/renovation/common/__tests__/ink_ripple.test.tsx
@@ -1,11 +1,11 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { h } from 'preact';
 import { shallow } from 'enzyme';
-import { InkRipple, viewFunction } from '../../js/renovation/ink-ripple';
-import { initConfig, showWave, hideWave } from '../../js/ui/widget/utils.ink_ripple';
+import { InkRipple, viewFunction } from '../ink_ripple';
+import { initConfig, showWave, hideWave } from '../../../ui/widget/utils.ink_ripple';
 
-jest.mock('../../js/ui/widget/utils.ink_ripple', () => ({
-  ...require.requireActual('../../js/ui/widget/utils.ink_ripple'),
+jest.mock('../../../ui/widget/utils.ink_ripple', () => ({
+  ...require.requireActual('../../../ui/widget/utils.ink_ripple'),
   initConfig: jest.fn(),
   showWave: jest.fn(),
   hideWave: jest.fn(),
@@ -30,6 +30,7 @@ describe('InkRipple', () => {
   describe('Behavior', () => {
     describe('Methods', () => {
       const event = {};
+
       describe('hideWave', () => {
         it('should call hide wave function', () => {
           new InkRipple({}).hideWave(event);
