+++ js/renovation/common/ink_ripple.tsx
@@ -1,7 +1,7 @@
 import {
   Component, ComponentBindings, JSXComponent, OneWay, Method,
 } from 'devextreme-generator/component_declaration/common';
-import { initConfig, showWave, hideWave } from '../ui/widget/utils.ink_ripple';
+import { initConfig, showWave, hideWave } from '../../ui/widget/utils.ink_ripple';
 
 // TODO: remake old ink ripple in new JSX component
 export const viewFunction = (model: InkRipple) => (
