+++ js/renovation/common/widget.tsx
@@ -10,17 +10,17 @@ import {
   Ref,
   Slot,
 } from 'devextreme-generator/component_declaration/common';
-import '../events/click';
-import '../events/hover';
+import '../../events/click';
+import '../../events/hover';
 
 import {
   active, dxClick, focus, hover, keyboard, resize, visibility,
-} from '../events/short';
-import { extend } from '../core/utils/extend';
-import { focusable } from '../ui/widget/selectors';
-import { isFakeClickEvent } from '../events/utils/index';
-import { normalizeStyleProp } from '../core/utils/style';
-import BaseWidgetProps from './utils/base-props';
+} from '../../events/short';
+import { extend } from '../../core/utils/extend';
+import { focusable } from '../../ui/widget/selectors';
+import { isFakeClickEvent } from '../../events/utils/index';
+import { normalizeStyleProp } from '../../core/utils/style';
+import BaseWidgetProps from '../utils/base_props';
 
 const getAria = (args): { [name: string]: string } => Object.keys(args).reduce((r, key) => {
   if (args[key]) {
