+++ js/renovation/list.tsx
@@ -3,11 +3,11 @@ import {
 } from 'devextreme-generator/component_declaration/common';
 /* eslint-disable import/named */
 import DataSource, { DataSourceOptions } from '../data/data_source';
-import { WidgetProps } from './widget';
+import { WidgetProps } from './common/widget';
 import DxList, { dxListItem, Options } from '../ui/list';
 import { dxElement } from '../core/element';
 import { event } from '../events/index';
-import renderTemplate from './utils/render-template';
+import renderTemplate from './utils/render_template';
 
 export const viewFunction = (viewModel: List) => (
   <div ref={viewModel.widgetRef as any} />
