+++ js/renovation/select_box.tsx
@@ -1,7 +1,7 @@
 import {
   Ref, Effect, Component, ComponentBindings, JSXComponent, Event, OneWay, TwoWay,
 } from 'devextreme-generator/component_declaration/common';
-import { WidgetProps } from './widget';
+import { WidgetProps } from './common/widget';
 // https://github.com/benmosher/eslint-plugin-import/issues/1699
 /* eslint-disable-next-line import/named */
 import DataSource, { DataSourceOptions } from '../data/data_source';
