+++ js/renovation/utils/base_props.ts
