+++ js/renovation/utils/render_template.ts
